+++ supabase-svelte/src/assets/json/apidoc.json
@@ -1,26 +1,15 @@
 {
-    "Globals" : {
-        "description" : "These are objects that are globally available. Some of them like Math and console provided by the browser. Others are custom utilities for the game.",
-        "fields" : [
-            "console", 
-            "Math", 
-            "Vector2D"
-        ],
-        "functions" : [
-            "H", 
-            "W",
-            "overlapCircle", 
-            "dist", 
-            "performance"
-        ]
-    },
-    "Game Object" : {
-        "description" : "These fields are shared by all game objects and are write-only.",
-        "fields" : [
-            "transform", 
-            "collider", 
-            "resources"
-        ],
-        "functions" : []
-    }
+    "Globals" : [
+        "console",
+        "Math",
+        "Vector2D",
+        "Graphics",
+        "Game"
+    ],
+    "Objects" : [
+        "Ship",
+        "Base",
+        "Asteroid",
+        "GameObject"
+    ]
 }
